requirements.txt
python-multipart
fastapi
uvicorn[standard]
torch
torchvision
pillow
einops
pytorch-lightning
prometheus-fastapi-instrumentator